fix bug

+++ src/single_baseline_main.py
@@ -90,7 +90,7 @@ def main():
         optimizer = torch.optim.Adam(global_model.parameters(), lr=args.lr,
                                      weight_decay=1e-4)
 
-    trainloader = DataLoader(train_dataset, batch_size=args.local_bs * (args.num_users * args.frac), shuffle=True, num_workers=args.workers,
+    trainloader = DataLoader(train_dataset, batch_size=int(args.local_bs * (args.num_users * args.frac)), shuffle=True, num_workers=args.workers,
                              pin_memory=use_cuda, drop_last=True)
     if args.dataset == 'cub200':
         criterion = torch.nn.CrossEntropyLoss().to(device)
