+++ src/single_combi_main.py
@@ -108,7 +108,7 @@ def main():
         optimizer = torch.optim.Adam(global_model.parameters(), lr=args.lr,
                                      weight_decay=1e-4)
 
-    trainloader = DataLoader(train_dataset, batch_size=args.local_bs * (args.num_users * args.frac), shuffle=True, num_workers=args.workers, pin_memory=use_cuda,
+    trainloader = DataLoader(train_dataset, batch_size=int(args.local_bs * (args.num_users * args.frac)), shuffle=True, num_workers=args.workers, pin_memory=use_cuda,
                              drop_last=True)
     criterion = torch.nn.NLLLoss().to(device)
     epoch_loss = []
